+++ src/main/java/org/opensearch/knn/profile/ProfileDefaultKNNWeight.java
@@ -15,7 +15,7 @@ import org.opensearch.knn.index.VectorDataType;
 import org.opensearch.knn.index.engine.KNNEngine;
 import org.opensearch.knn.index.memory.NativeMemoryAllocation;
 import org.opensearch.knn.index.query.DefaultKNNWeight;
-import org.opensearch.knn.index.query.ExactSearcher;
+import org.opensearch.knn.index.query.exactsearch.ExactSearcher;
 import org.opensearch.knn.index.query.KNNQuery;
 import org.opensearch.knn.index.query.SegmentLevelQuantizationInfo;
 import org.opensearch.knn.profile.query.KNNMetrics;